add (profile, app, auth components) final uploads

+++ src/app/components/auth/auth.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core'
 import { HttpClient, HttpHeaders } from '@angular/common/http'
 import { Router } from "@angular/router"
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-auth',
@@ -28,22 +29,33 @@ export class AuthComponent implements OnInit {
       session: null
     }
 
-    constructor(private http: HttpClient, private router: Router) {
+    constructor(private http: HttpClient, private router: Router) {}
+
+    ngOnInit() {
       if (localStorage.getItem('currentUser') !== null) {
         this.location('/profile')
       }
       document.getElementsByTagName('body')[0].classList.add("bodyman")
     }
 
-    ngOnInit() { }
-
     authUser(): boolean {
       this.http.post('http://localhost:88/auth', this.authFormModel)
       .subscribe(
         res => {
-          this.user = res
-          localStorage.setItem('currentUser', JSON.stringify(this.user));
-          this.router.navigate(['/profile'])
+          let message: any = res
+          if (message.type === 'error') {
+            Swal.fire({
+              title: 'Ошибка',
+              text: message.message,
+              icon: 'error',
+              showCancelButton: false,
+              confirmButtonText: 'Закрыть',
+            })
+          } else {
+            this.user = res
+            localStorage.setItem('currentUser', JSON.stringify(this.user));
+            this.router.navigate(['/profile'])
+          }
         },
         err => {
           console.error(err)
@@ -56,9 +68,20 @@ export class AuthComponent implements OnInit {
       this.http.post('http://localhost:88/reg', this.regFormModel)
       .subscribe(
         res => {
-          this.user = res
-          localStorage.setItem('currentUser', JSON.stringify(this.user));
-          this.router.navigate(['/profile'])
+          let message: any = res
+          if (message.type = "error") {
+            Swal.fire({
+              title: 'Ошибка',
+              text: message.message,
+              icon: 'error',
+              showCancelButton: false,
+              confirmButtonText: 'Закрыть',
+            })
+          } else {
+            this.user = res
+            localStorage.setItem('currentUser', JSON.stringify(this.user));
+            this.router.navigate(['/profile'])
+          }
         },
         err => {
           console.log(err)
@@ -84,4 +107,5 @@ export class AuthComponent implements OnInit {
     location(url: string): void {
       this.router.navigate([url]);
     }
+
 }
